Removes unused prop
