+++ pages/applications/unassigned.tsx
@@ -1,7 +1,9 @@
+import React, { useState, SyntheticEvent } from 'react';
 import { GetServerSideProps } from 'next';
 import { useRouter } from 'next/router';
-import React, { useState, SyntheticEvent } from 'react';
-import ApplicationTable from '../../components/admin/application-table';
+import ApplicationsTable from '../../components/admin/ApplicationsTable';
+import SimplePaginationSearch from '../../components/SimplePaginationSearch';
+
 import {
   HorizontalNav,
   HorizontalNavItem,
@@ -13,7 +15,7 @@ import Layout from '../../components/layout/staff-layout';
 import { HackneyGoogleUser } from '../../domain/HackneyGoogleUser';
 import {
   APPLICATION_UNNASIGNED,
-  PaginatedApplicationListResponse,
+  PaginatedSearchResultsResponse,
 } from '../../domain/HousingApi';
 import { UserContext } from '../../lib/contexts/user-context';
 import {
@@ -24,7 +26,7 @@ import { getRedirect, getSession } from '../../lib/utils/googleAuth';
 
 interface PageProps {
   user?: HackneyGoogleUser;
-  applications: PaginatedApplicationListResponse | null;
+  applications: PaginatedSearchResultsResponse | null;
 }
 
 export default function ApplicationListPage({
@@ -43,13 +45,6 @@ export default function ApplicationListPage({
     });
 
     setActiveNavItem(name);
-  };
-
-  const setPaginationToken = (paginationToken: string | null) => {
-    router.push({
-      pathname: router.pathname,
-      query: { ...router.query, paginationToken },
-    });
   };
 
   return (
@@ -83,14 +78,20 @@ export default function ApplicationListPage({
                 Reviews
               </HorizontalNavItem>
             </HorizontalNav>
-            <ApplicationTable
-              applications={applications}
-              initialPaginationToken={
-                router.query.paginationToken as string | undefined
-              }
-              setPaginationToken={setPaginationToken}
-              showStatus={false}
-            />
+            {applications ? (
+              <>
+                <ApplicationsTable
+                  applications={applications}
+                  showStatus={true}
+                />
+
+                <SimplePaginationSearch
+                  totalItems={applications.totalResults}
+                  page={applications.page}
+                  numberOfItemsPerPage={applications.pageSize}
+                />
+              </>
+            ) : null}
           </div>
         </div>
       </Layout>
